PEPPER-1171 Removed KitDao interface, replacing it with KitDaoImpl.
