+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/db/dao/mercury/MercurySampleDao.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.dsm.db.dao.Dao;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.db.dto.mercury.MercurySampleDto;
 import org.broadinstitute.dsm.statics.DBConstants;
 import org.broadinstitute.lddp.db.SimpleResult;
@@ -79,12 +79,12 @@ public class MercurySampleDao implements Dao<MercurySampleDto> {
     }
 
     public List<MercurySampleDto> findEligibleSamples(String ddpParticipantId, String realm,
-                                                      KitDaoImpl kitDaoImpl) {
+                                                      KitDao kitDao) {
         ArrayList<MercurySampleDto> samples = new ArrayList<>();
         SimpleResult results = inTransaction((conn) -> {
             SimpleResult dbVals = new SimpleResult();
             try {
-                samples.addAll(getEligibleTissues(conn, ddpParticipantId, realm, kitDaoImpl));
+                samples.addAll(getEligibleTissues(conn, ddpParticipantId, realm, kitDao));
                 samples.addAll(getEligibleKits(conn, ddpParticipantId, realm));
             } catch (Exception e) {
                 dbVals.resultException = e;
@@ -130,9 +130,9 @@ public class MercurySampleDao implements Dao<MercurySampleDto> {
     }
 
     private List<MercurySampleDto> getEligibleTissues(Connection conn, String ddpParticipantId, String realm,
-                                                      KitDaoImpl kitDaoImpl) throws Exception {
+                                                      KitDao kitDao) throws Exception {
         ArrayList<MercurySampleDto> samples = new ArrayList<>();
-        if (!participantHasKitReceived(conn, ddpParticipantId, kitDaoImpl)) {
+        if (!participantHasKitReceived(conn, ddpParticipantId, kitDao)) {
             return samples;
         }
         SimpleResult dbVals = new SimpleResult();
@@ -164,8 +164,8 @@ public class MercurySampleDao implements Dao<MercurySampleDto> {
         return samples;
     }
 
-    private boolean participantHasKitReceived(Connection conn, String ddpParticipantId, KitDaoImpl kitDaoImpl) {
-        return kitDaoImpl.hasKitReceived(conn, ddpParticipantId);
+    private boolean participantHasKitReceived(Connection conn, String ddpParticipantId, KitDao kitDao) {
+        return kitDao.hasKitReceived(conn, ddpParticipantId);
     }
 
     private String getSampleStatus(ResultSet rs) throws SQLException {
