+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/route/kit/KitFinalScanRoute.java
@@ -3,7 +3,7 @@ package org.broadinstitute.dsm.route.kit;
 import java.util.List;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.model.kit.KitFinalScanUseCase;
 import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 
@@ -14,7 +14,7 @@ public class KitFinalScanRoute extends KitStatusChangeRoute {
 
     @Override
     protected void processRequest() {
-        KitFinalScanUseCase kitFinalScanUseCase = new KitFinalScanUseCase(kitPayload, new KitDaoImpl());
+        KitFinalScanUseCase kitFinalScanUseCase = new KitFinalScanUseCase(kitPayload, new KitDao());
         scanErrorList.addAll(kitFinalScanUseCase.get());
     }
