+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/route/kit/KitTrackingScanRoute.java
@@ -3,7 +3,7 @@ package org.broadinstitute.dsm.route.kit;
 import java.util.List;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.model.kit.KitTrackingScanUseCase;
 import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 
@@ -14,7 +14,7 @@ public class KitTrackingScanRoute extends KitStatusChangeRoute {
 
     @Override
     protected void processRequest() {
-        KitTrackingScanUseCase kitTrackingScanUseCase = new KitTrackingScanUseCase(kitPayload, new KitDaoImpl());
+        KitTrackingScanUseCase kitTrackingScanUseCase = new KitTrackingScanUseCase(kitPayload, new KitDao());
         scanErrorList.addAll(kitTrackingScanUseCase.get());
     }
