+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/route/kit/ReceivedKitsRoute.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.NonNull;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.model.kit.KitReceivedUseCase;
 import org.broadinstitute.dsm.util.NotificationUtil;
 import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
@@ -16,7 +16,7 @@ public class ReceivedKitsRoute extends KitStatusChangeRoute {
 
     @Override
     protected void processRequest() {
-        KitReceivedUseCase kitReceivedUseCase = new KitReceivedUseCase(kitPayload, new KitDaoImpl(), notificationUtil);
+        KitReceivedUseCase kitReceivedUseCase = new KitReceivedUseCase(kitPayload, new KitDao(), notificationUtil);
         scanErrorList.addAll(kitReceivedUseCase.get());
     }
