+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/route/kit/SentKitRoute.java
@@ -3,7 +3,7 @@ package org.broadinstitute.dsm.route.kit;
 import java.util.List;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.model.kit.KitSentUseCase;
 import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 
@@ -15,7 +15,7 @@ public class SentKitRoute extends KitStatusChangeRoute {
 
     @Override
     protected void processRequest() {
-        KitSentUseCase kitSentUseCase = new KitSentUseCase(kitPayload, new KitDaoImpl());
+        KitSentUseCase kitSentUseCase = new KitSentUseCase(kitPayload, new KitDao());
         scanErrorList.addAll(kitSentUseCase.get());
     }
