+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/route/mercury/GetMercuryEligibleSamplesRoute.java
@@ -3,7 +3,7 @@ package org.broadinstitute.dsm.route.mercury;
 import java.util.List;
 
 import lombok.extern.slf4j.Slf4j;
-import org.broadinstitute.dsm.db.dao.kit.KitDaoImpl;
+import org.broadinstitute.dsm.db.dao.kit.KitDao;
 import org.broadinstitute.dsm.db.dao.mercury.MercurySampleDao;
 import org.broadinstitute.dsm.db.dto.mercury.MercurySampleDto;
 import org.broadinstitute.dsm.security.RequestHandler;
@@ -17,17 +17,17 @@ import spark.Response;
 @Slf4j
 public class GetMercuryEligibleSamplesRoute extends RequestHandler {
     private MercurySampleDao mercurySampleDao;
-    private KitDaoImpl kitDaoImpl;
+    private KitDao kitDao;
     public static String projectId;
     public static String topicId;
 
 
     public GetMercuryEligibleSamplesRoute(MercurySampleDao mercurySampleDao, String projectId, String topicId,
-                                          KitDaoImpl kitDao) {
+                                          KitDao kitDao) {
         this.mercurySampleDao = mercurySampleDao;
         this.projectId = projectId;
         this.topicId = topicId;
-        this.kitDaoImpl = kitDao;
+        this.kitDao = kitDao;
     }
 
     @Override
@@ -48,7 +48,7 @@ public class GetMercuryEligibleSamplesRoute extends RequestHandler {
             throw new RuntimeException("No ddpParticipantId query param was sent");
         }
         String ddpParticipantId = queryParams.get(RoutePath.DDP_PARTICIPANT_ID).value();
-        List<MercurySampleDto> eligibleSamples = mercurySampleDao.findEligibleSamples(ddpParticipantId, realm, kitDaoImpl);
+        List<MercurySampleDto> eligibleSamples = mercurySampleDao.findEligibleSamples(ddpParticipantId, realm, kitDao);
         log.info(String.format("Returning a list of %d samples", eligibleSamples.size()));
         return eligibleSamples;
     }
